Added encrypt to handler.
